Moving old alamopy to alamopy_depr and fixing all imports I could find
