+++ idaes/surrogate/alamopy_depr/remapminmax.py
